Finish code part of task4

+++ src/test/java/task3/SubTaskA.java
@@ -15,20 +15,20 @@ public class SubTaskA extends DriverSetup {
         softAssert.assertEquals(homePage.getLoggedName(), ExpectedData.LOGGED_USER_NAME);
 
         // 5. Assert that there are 4 items on the header section are displayed , and they have proper texts
-        softAssert.assertEquals(homePage.getHeaderSection().getItemsSize(), ExpectedData.HEADER_SECTION_ITEMS.size());
+        softAssert.assertEquals((int) homePage.getHeaderSection().getItemsSize(), (int) ExpectedData.HEADER_SECTION_ITEMS.size());
         softAssert.assertEquals(homePage.getHeaderSection().getItemsNames(), ExpectedData.HEADER_SECTION_ITEMS);
         homePage.getHeaderSection().getItems().forEach(item -> {
             softAssert.assertTrue(item.isDisplayed());
         });
 
         // 6. Assert that there are 4 images on the Index Page, and they are displayed
-        softAssert.assertEquals(homePage.getBenefitIconsSize(), ExpectedData.BENEFIT_ICONS.size());
+        softAssert.assertEquals((int) homePage.getBenefitIconsSize(), (int) ExpectedData.BENEFIT_ICONS.size());
         homePage.getBenefitIcons().forEach(icon -> {
             softAssert.assertTrue(icon.isDisplayed());
         });
 
         // 7. Assert that there are 4 texts on the Index Page under icons, and they have proper text
-        softAssert.assertEquals(homePage.getBenefitIconsTextsSize(), ExpectedData.BENEFIT_ICONS.size());
+        softAssert.assertEquals((int) homePage.getBenefitIconsTextsSize(), (int) ExpectedData.BENEFIT_ICONS.size());
         softAssert.assertEquals(homePage.getBenefitIconsTextsStrings(), ExpectedData.BENEFIT_ICONS);
 
         // 8. Assert that there is an iframe with frame button exists
@@ -41,7 +41,7 @@ public class SubTaskA extends DriverSetup {
         homePage.switchToOriginalWindow();
 
         // 11. Assert that there are 5 items in the Left Section are displayed, and they have proper text
-        softAssert.assertEquals(homePage.getLeftSection().getItemSize(), ExpectedData.LEFT_SECTION_ITEMS.size());
+        softAssert.assertEquals((int) homePage.getLeftSection().getItemSize(), (int) ExpectedData.LEFT_SECTION_ITEMS.size());
         homePage.getLeftSection().getItems().forEach(item -> {
             softAssert.assertTrue(item.isDisplayed());
         });
